Refactor: 동적 라우팅을 위한 폴더 이름 수정 및 코드 수정
